Added styling to login page
