+++ frontend/src/components/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { AuthRoute } from "../util/route_util";
 import { Switch } from "react-router-dom";
-import NavBarContainer from "./nav/navbar_container";
+// import NavBarContainer from "./nav/navbar_container";
 
 import MainPage from "./main/main_page";
 import LoginFormContainer from "./session/login_form_container";
@@ -9,7 +9,7 @@ import SignupFormContainer from "./session/signup_form_container";
 
 const App = () => (
   <div>
-    <NavBarContainer />
+    {/* <NavBarContainer /> */}
     <Switch>
       <AuthRoute exact path="/" component={MainPage} />
       <AuthRoute exact path="/login" component={LoginFormContainer} />
