+++ frontend/src/components/nav/navbar_container.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import { logout } from "../../actions/session_actions";
 
-import NavBar from "./navbar";
+import NavBar from "./navbar.jsx";
 
 const mapStateToProps = state => ({
   loggedIn: state.session.isAuthenticated
