+++ frontend/src/components/session/login_form_container.js
@@ -1,6 +1,7 @@
 import { connect } from "react-redux";
 import { login } from "../../actions/session_actions";
-import LoginForm from "./login_form";
+import LoginForm from "./login_form.jsx";
+import "./login_form.css";
 
 const mapStateToProps = state => {
   return {
