+++ frontend/src/components/session/signup_form_container.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
 import { signup } from "../../actions/session_actions";
-import SignupForm from "./signup_form";
+import SignupForm from "./signup_form.jsx";
 
 const mapStateToProps = state => {
   return {
